+++ CHANGELOG.md
@@ -9,6 +9,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Sprint 5
 
+### [0.0.21] - 2025-10-24
+
+### Refactor
+
+- Test backgammon into test bar,basic,bearing_off,checkers,moves
+
+### Docs
+
+- Complete justification.md
+- Complete docstring with args and return
+
+#### Added
+
+- attribute encapsulation
+
+### Fix
+
+- cli test fail
+
 ### [0.0.21] - 2025-10-23
 
 #### Added